requirements.txt
Flask==3.1.2
matplotlib==3.10.0
numpy==2.4.1
python-dotenv
mistralai
